cleancode问题修改
# message auto-generated for no-merge-commit merge:
merge timeline_device_feat_****** into master

cleancode问题修改

Created-by: insightDeveloper

Match-id-300fa5267211a266414c9c2762e30db6501df7ab

+++ server/src/modules/full_db/ProjectParse/ProjectParserDb.cpp
@@ -223,7 +223,7 @@ ProjectTypeEnum ProjectParserDb::GetProjectType(const std::string &dataPath)
     }
     std::vector<std::string> msprofFiles = FileUtil::FindFilesWithFilter(dataPath, std::regex(msprofDBReg));
     std::vector<std::string> clusterPath = FileUtil::FindFilesWithFilter(dataPath, std::regex(clusterDBReg));
-    int rankCount = frameworkFiles.size() + msprofFiles.size();
+    uint64_t rankCount = frameworkFiles.size() + msprofFiles.size();
     // 如果rank的数据大于1个或导入的为cluster_analysis.db单文件，则判断需要进行集群分析
     bool isCluster = (rankCount > 1) || (rankCount == 0 && (clusterPath.size() > 0));
     return isCluster ? ProjectTypeEnum::DB_CLUSTER : ProjectTypeEnum::DB;
